+++ lib/tucan.ex
@@ -344,6 +344,7 @@ defmodule Tucan do
 
   ## Plots
 
+  # TODO: move it to helper module for reusability
   # global_opts should be applicable in all plot types
   @global_opts [:width, :height, :title, :only]
   @global_mark_opts [:clip, :fill_opacity, :tooltip]